+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- envs: Parameter to disable frequency checks during frequency regulation
 - PID balancer: Monitor both tire contacts in simulation examples
 
 ### Changed